Add group members

+++ src/structures/GroupChannel.ts
@@ -1,7 +1,8 @@
 import { GroupChannel as RawGroupChannel } from 'revolt-api/types/Channels'
 import { TextBasedChannel } from './interfaces/TextBasedChannel'
 import { Client, User, UserResolvable } from '..'
-import { ChannelPermissions, ChannelTypes } from '../util'
+import { TypeError } from '../errors'
+import { ChannelPermissions, ChannelTypes, Collection } from '../util'
 
 export class GroupChannel extends TextBasedChannel {
     name!: string
@@ -9,14 +10,22 @@ export class GroupChannel extends TextBasedChannel {
     ownerId!: string
     readonly type = ChannelTypes.GROUP
     permissions!: Readonly<ChannelPermissions>
+    icon: string | null = null
+    _users: string[] = []
     constructor(client: Client, data: RawGroupChannel) {
         super(client, data)
         this._patch(data)
     }
 
     _patch(data: RawGroupChannel): this {
+        if (!data) return this
+
         if ('description' in data) {
             this.description = data.description ?? null
+        }
+
+        if (Array.isArray(data.recipients)) {
+            this._users = data.recipients
         }
 
         if (typeof data.permissions === 'number') {
@@ -25,6 +34,10 @@ export class GroupChannel extends TextBasedChannel {
 
         if (data.owner) {
             this.ownerId = data.owner
+        }
+
+        if ('icon' in data) {
+            this.icon = data.icon?._id ?? null
         }
 
         if (data.name) {
@@ -40,21 +53,43 @@ export class GroupChannel extends TextBasedChannel {
         return clone
     }
 
-    get owner(): User | null {
-        return this.client.users.cache.get(this.ownerId) ?? null
-    }
-
     async add(user: UserResolvable): Promise<void> {
         const userId = this.client.users.resolveId(user)
+        if (!userId) throw new TypeError('INVALID_TYPE', 'user', 'UserResolvable')
         await this.client.api.put(`/channels/${this.id}/recipients/${userId}`)
     }
 
     async remove(user: UserResolvable): Promise<void> {
         const userId = this.client.users.resolveId(user)
+        if (!userId) throw new TypeError('INVALID_TYPE', 'user', 'UserResolvable')
         await this.client.api.delete(`/channels/${this.id}/recipients/${userId}`)
     }
 
     async leave(): Promise<void> {
         await super.delete()
+    }
+
+    iconURL(options?: { size: number }): string | null {
+        if (!this.icon) return null
+        return this.client.endpoints.icon(this.icon, options?.size)
+    }
+
+    get owner(): User | null {
+        return this.client.users.cache.get(this.ownerId) ?? null
+    }
+
+    get users(): Collection<string, User> {
+        const users = new Collection<string, User>()
+
+        for (const userId of this._users) {
+            const user = this.client.users.cache.get(userId)
+            if (user) users.set(user.id, user)
+        }
+
+        return users
+    }
+
+    get members(): Collection<string, User> {
+        return this.users
     }
 }
